Create objTest.js
This project is in progress. It takes the players first entry and gives a result to proceed to the next player input. This first guess is yet to be must be removed and the generated number needs to be stored for proceeding guesses, until the player is correct. This is to follow.